requirements.txt
requests
matplotlib
python-dotenv
types-requests
